[ops] move !debug and !debug-vd into ops/debug.py
