+++ aipl/ops/op_misc.py
@@ -11,34 +11,6 @@ class UserAbort(BaseException):
 def op_abort(aipl, *args):
     raise UserAbort(*args)
 
-
-@defop('debug', None, None, arity=0)
-def op_debug(aipl, *args):
-    aipl.debug = True
-    aipl.single_step = lambda *args, **kwargs: breakpoint()
-
-
-def vd_singlestep(inputs:List[LazyRow], cmd):
-    import visidata
-    @visidata.VisiData.api
-    def uberquit(vd):
-        raise UserAbort('user abort')
-
-    inputs = list(r._asdict() for r in inputs)
-    sheet = visidata.PyobjSheet('current_input', source=inputs)
-    sheet.help = '{sheet.recentcmd}'
-    argstr = ' '.join(str(x) for x in cmd.args)
-    kwargstr = ' '.join(f'{k}={v}' for k, v in cmd.kwargs.items())
-    sheet.recentcmd = f'[line {cmd.linenum}] !' + ' '.join([cmd.opname, argstr, kwargstr])
-    sheet.addCommand('Q', 'quit-really', 'uberquit()')
-    visidata.vd.run(sheet)
-
-
-@defop('debug-vd', None, None, arity=0)
-def op_debug_vd(aipl):
-    aipl.single_step = vd_singlestep
-
-
 @defop('assert-equal', 0, None)
 def op_assert_equal(aipl, v:str, prompt=''):
     if v != prompt:
